refactor(op-rbuilder): move execution tests to separate file
Convert the execution module from a single file into a directory module
so its unit tests live in execution/tests.rs instead of an inline
#[cfg(test)] block, per review feedback on #21250.
